# public/openapi.yaml
openapi: 3.0.0
info:
  title: Beantown Baby Diaper Bank API
  version: 1.0.0
  description: OpenAPI docs aligned with implemented Next.js API routes.
servers:
  - url: /api
    description: Internal API
paths:
...
  /distributions:
    get:
      summary: Get distributions
      parameters:
        - name: month
          in: query
          schema:
            $ref: '#/components/schemas/MonthEnum'
        - name: year
          in: query
          schema:
            type: string
        - name: startMonth
          in: query
          schema:
            $ref: '#/components/schemas/MonthEnum'
        - name: startYear
          in: query
          schema:
            type: string
        - name: endMonth
          in: query
          schema:
            $ref: '#/components/schemas/MonthEnum'
        - name: endYear
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Distribution records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DistributionItem'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Delete distributions by IDs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids]
              properties:
                ids:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Deletion count
          content:
            application/json:
              schema:
                type: object
                properties:
                  deletedCount:
                    type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    MonthEnum:
      type: string
      enum:
        - January
        - February
        - March
        - April
        - May
        - June
        - July
        - August
        - September
        - October
        - November
        - December
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
...
    DistributionItem:
      type: object
      properties:
        id:
          type: string
        createdAt:
          type: string
          format: date-time
        partnerId:
          type: string
          nullable: true
        cityId:
          type: string
          nullable: true
        year:
          type: string
          nullable: true
        month:
          type: string
          nullable: true
        numberDiapers:
          type: string
          nullable: true
        numberChildren:
          type: string
          nullable: true
        percentage:
          type: number
          nullable: true
        partner:
          type: object
          nullable: true
          properties:
            name:
              type: string
              nullable: true
        city:
          type: object
          nullable: true
          properties:
            name:
              type: string
              nullable: true
...